+++ flycheck.el
@@ -3542,14 +3542,27 @@ the beginning of the buffer."
     ;; list itself.
     (flycheck-error-list-set-source (current-buffer))))
 
+(define-button-type 'flycheck-error-list
+  'action #'flycheck-error-list-button-goto-error
+  'help-echo (purecopy "mouse-2, RET: goto error"))
+
+(defun flycheck-error-list-button-goto-error (button)
+  "Go to the error at BUTTON."
+  (flycheck-error-list-goto-error (button-start button)))
+
+(defun flycheck-error-list-make-cell (text &optional face)
+  "Make an error list cell with TEXT and FACE."
+  (let ((face (or face 'default)))
+    (list text 'type 'flycheck-error-list 'face face)))
+
 (defun flycheck-error-list-make-number-cell (number face)
   "Make a table cell for a NUMBER with FACE.
 
 Convert NUMBER to string, fontify it with FACE and return the
 string with attached text properties."
-  (if (numberp number)
-      (propertize (number-to-string number) 'font-lock-face face)
-    ""))
+  (flycheck-error-list-make-cell
+   (if (numberp number) (number-to-string number) "")
+   face))
 
 (defun flycheck-error-list-make-entry (error)
   "Make a table cell for the given ERROR.
@@ -3566,9 +3579,10 @@ Return a list with the contents of the table cell."
                    line 'flycheck-error-list-line-number)
                   (flycheck-error-list-make-number-cell
                    column 'flycheck-error-list-column-number)
-                  (propertize (symbol-name (flycheck-error-level error))
-                              'font-lock-face level-face)
-                  (format "%s (%s)" message checker)))))
+                  (flycheck-error-list-make-cell
+                   (symbol-name (flycheck-error-level error)) level-face)
+                  (flycheck-error-list-make-cell
+                   (format "%s (%s)" message checker))))))
 
 (defun flycheck-error-list-entries ()
   "Create the entries for the error list."
@@ -3658,13 +3672,12 @@ list."
       (revert-buffer))
     (flycheck-error-list-highlight-errors)))
 
-(defun flycheck-error-list-goto-error (&optional error)
-  "Go to the location of an ERROR from the error list.
+(defun flycheck-error-list-goto-error (&optional pos)
+  "Go to the location of the error at POS in the error list.
 
-Interactively or when ERROR is not given, go to the error at
-point."
+POS defaults to `point'."
   (interactive)
-  (-when-let* ((error (or error (tabulated-list-get-id)))
+  (-when-let* ((error (tabulated-list-get-id pos))
                (buffer (flycheck-error-buffer error)))
     (when (buffer-live-p buffer)
       (if (eq (window-buffer) (get-buffer flycheck-error-list-buffer))